Rename index to downloadTrack
